requirements.txt
# BEGIN: gcp_data_ingestion.py
spotipy==2.19.0
python-dotenv==0.19.1
pandas==1.3.3
google-cloud-bigquery==2.26.0
pandas_gbq==0.15.0
tqdm==4.62.3
# END: gcp_data_ingestion.py